requirements.txt
numpy~=1.23.4
torch~=1.13.1
torchdiffeq~=0.2.3
tqdm~=4.64.1
torchvision~=0.14.1
PyYAML~=6.0
easydict~=1.10
tensorboard~=2.11.0